Release 2.2 from develop

+++ src/main/resources/static/js/incidents/incident-grid-service.js
@@ -34,6 +34,10 @@ function IncidentGridService() {
             });
     }
 
+    this.generateExcel = () => {
+        window.location.href = `/reports/incidents/excel?from=${this.filterFrom}&to=${this.filterTo}`;
+    }
+
     this.generateReport = () => {
         fetch(`/reports/incidents?from=${this.filterFrom}&to=${this.filterTo}`)
             .then(response => {
@@ -125,7 +129,6 @@ function IncidentGridService() {
         this.incidentGrid.render(document.getElementById("incidentsTable"));
         searchService.initSearch(this.incidentGrid, this.currentConfig);
         gridOptions.init(this.incidentGrid, document.getElementById("gridOptions"));
-
     }
 
     this.mapRow = (field) => {
@@ -158,11 +161,13 @@ function IncidentGridService() {
                 formatter: (cell, row) => {
                     const url = viewUrl + row.cells[0]['data'];
                     return gridjs.html(`<a href="${url}">${cell}</a>`);
-                }
+                },
+                width: '250px'
             },
             {
                 id: "createdAt",
-                name: "Oprettet"
+                name: "Oprettet",
+                width: '120px'
             }];
         columnNames.forEach(c => {
             this.columns.push({
@@ -184,9 +189,11 @@ function IncidentGridService() {
                 formatter: (cell, row) => {
                     const id = row.cells[0]['data'];
                     const name = row.cells[1]['data'];
-                    return gridjs.html(
-                        `<button type="button" class="btn btn-icon btn-outline-light btn-xs me-1" onclick="incidentService.editIncident('editIncidentDialog', '${id}')"><i class="pli-pencil fs-5"></i></button>`
-                        + `<button type="button" class="btn btn-icon btn-outline-light btn-xs me-1" onclick="incidentService.deleteIncident(incidentGridService.incidentGrid, '${id}', '${name}')"><i class="pli-trash fs-5"></i></button>`);
+                    if(superuser) {
+                        return gridjs.html(
+                            `<button type="button" class="btn btn-icon btn-outline-light btn-xs me-1" onclick="incidentService.editIncident('editIncidentDialog', '${id}')"><i class="pli-pencil fs-5"></i></button>`
+                            + `<button type="button" class="btn btn-icon btn-outline-light btn-xs me-1" onclick="incidentService.deleteIncident(incidentGridService.incidentGrid, '${id}', '${name}')"><i class="pli-trash fs-5"></i></button>`);
+                    }
                 }
             });
     }
